requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.4.1
Babel==2.10.1
click==8.0.1
fastapi==0.110.1
flake8==4.0.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httptools==0.2.0
idna==3.3
importlib-metadata==4.2.0
importlib-resources==5.7.1
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.6.1
py-moneyed==2.0
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
PyMySQL==1.0.2
pytest==8.1.1
python-dotenv==0.19.0
pytz==2022.1
PyJWT==2.8.0
PyYAML==5.4.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
SQLAlchemy-Utils==0.38.2
starlette==0.37.2
typing-extensions==4.8.0
uvicorn==0.15.0
uvloop==0.16.0
validate-cpf==0.1.2
watchgod==0.7
websockets==10.0
zipp==3.6.0
